Modernize LSP setup: Remove ALE conflicts, switch to Mason
- Removed ALE plugin completely to avoid diagnostic conflicts
- Moved from manual LSP setup to modern Mason-based management
- Created proper plugins/lsp.lua with Mason + mason-lspconfig
- Mason auto-installs: ts_ls, ruby_lsp, pyright, lua_ls
- Cleaned up init.lua to only contain user's original config
- Updated setup script to reflect Mason workflow
- No more manual language server installation needed

Amp-Thread-ID: https://ampcode.com/threads/T-75cc897e-2331-4946-a1fc-645f384079fa
